Added source_attribute pattern, ReferenceEntry NamedConfiguration, split specs into per-class files - resolves #986
