+++ spec/models/option_configs/extra_option_configs/label_spec.rb
@@ -1,0 +1,83 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+require './db/table_generators/dynamic_models_table'
+
+# Tests for Label configuration class.
+# Verifies configure_direct with type :string, store_processed_value?,
+# and prepare_config defaulting to humanized name from parent.
+RSpec.describe 'ExtraOptionConfigs::Label', type: :model do
+  include MasterSupport
+  include ModelSupport
+  include DynamicModelSupport
+  include ExtraOptionConfigsSupport
+
+  before(:each) do
+    create_admin
+    create_user
+    setup_access :trackers
+    setup_access :tracker_histories
+    @dm = generate_test_dynamic_model
+    setup_access :dynamic_model__test_created_by_recs, user: @user
+  end
+
+  let(:klass) { OptionConfigs::ExtraOptionConfigs::Label }
+
+  describe 'class structure' do
+    it 'inherits from ExtraOptionConfigs::BaseConfiguration' do
+      expect(klass.ancestors).to include(OptionConfigs::ExtraOptionConfigs::BaseConfiguration)
+    end
+
+    it 'does not inherit from ConfigBase' do
+      expect(klass.ancestors).not_to include(OptionConfigs::ExtraOptionConfigs::ConfigBase)
+    end
+
+    it 'declares configure_direct with type :string' do
+      expect(klass.option_types[:direct]).to include(:label)
+    end
+  end
+
+  describe 'initialization' do
+    it 'stores the string as label attribute' do
+      instance = klass.new('My Label')
+      expect(instance.label).to eq 'My Label'
+    end
+
+    it 'defaults to empty string when initialized with nil' do
+      instance = klass.new(nil)
+      expect(instance.label).to eq ''
+    end
+
+    it 'uses prepare_config to default to humanized name from parent' do
+      expect(klass).to respond_to(:prepare_config)
+    end
+  end
+
+  describe 'ExtraOptions integration' do
+    it 'defaults label to a humanized version of the config name' do
+      eo = config_for(<<~YAML)
+        my_custom_name:
+          fields:
+            - test1
+      YAML
+      expect(eo.label).to eq 'My custom name'
+    end
+
+    it 'preserves an explicitly set label' do
+      eo = config_for(<<~YAML)
+        default:
+          label: Custom Label
+      YAML
+      expect(eo.label).to eq 'Custom Label'
+    end
+
+    it 'defaults label to humanized name when not specified' do
+      eo = config_for(<<~YAML)
+        default:
+          fields:
+            - test1
+      YAML
+      expect(eo.label).to eq 'Default'
+    end
+  end
+end
